% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.bias.R
\name{CV}
\alias{CV}
\title{k flod cross valiation}
\usage{
CV(y, fold = 10)
}
\arguments{
\item{y}{The label vector, positive(1) or negative(0)}

\item{fold}{The number of fold of cross validation used in the program.}
}
\description{
k flod cross valiation
}
\examples{
result <- CV(c(1,0,1,1,1,1,1,1,0,0,0,1,0,0,0,1,0),4)
print(result)
}
