\name{FSCUDA-package}
\alias{FSCUDA}
\alias{FesSelectionCUDA}
\docType{package}
\title{
A gene feature selection algorithm.
}
\description{
There is a features selection algorithm to solve gene data which has a "big p small n" feature. Particularly, it selects subset which has both low feature numbers and high accuracies .Logistic regression classification algorithm is used for measuring chosen feature subsets.
}
\details{
\tabular{ll}{
Package: \tab FSCUDA\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-09-25\cr
License: \tab GPL-2\cr
}
}

\author{
Ruiquan Ge,Qinghan Meng ,Goku
Maintainer:Goku<jym775096422@gmail.com>
}


\keyword{ gene, features, selection, CUDA }


