% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGPU.R
\name{generateIds}
\alias{generateIds}
\title{generate GPU ids from the number of GPUs}
\usage{
generateIds(count)
}
\arguments{
\item{count}{The number of GPUs.}
}
\description{
generate GPU ids from the number of GPUs
}
\examples{
generateIds(3)
}
