% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGPU.R
\name{getGPUCount}
\alias{getGPUCount}
\title{get GPU count}
\usage{
getGPUCount(cl = NULL)
}
\arguments{
\item{cl}{The cluster of computers which you created. If it is NULL, the program will create clusters based on the number of devices automatically.}
}
\description{
get GPU count
}
\examples{
getGPUCount()
}
