% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGPU.R
\name{getGPUIds}
\alias{getGPUIds}
\title{get GPU ids}
\usage{
getGPUIds(cl = NULL)
}
\arguments{
\item{cl}{The cluster of computers which you created. If it is NULL, the program will create clusters based on the number of devices automatically.}
}
\description{
get GPU ids
}
\examples{
getGPUIds()
}
