% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.bias.R
\name{LOOCV}
\alias{LOOCV}
\title{loo cross valiation}
\usage{
LOOCV(features, X, Y)
}
\arguments{
\item{features}{The number of features}

\item{X}{Gene expression profile matrix. The columns represent different prob sets. The rows represent different samples. The values in the matrix represent gene expression levels.}

\item{Y}{The label vector, positive(1) or negative(0)}
}
\description{
loo cross valiation
}
\examples{
X =  matrix(1:12, 4, 3)
Y = c(1,0,1,1)
result <- LOOCV(3,X,Y)
print(result)
}
