% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRCUDA.R
\name{LRCUDA}
\alias{LRCUDA}
\title{Logistic Regression wiht CUDA}
\usage{
LRCUDA(x, y, n.comb = 2, error.threshhold = 0, fold = 10, device.id = 0,
  cl = NULL)
}
\arguments{
\item{x}{Gene expression profile matrix. The columns represent different prob sets. The rows represent different samples. The values in the matrix represent gene expression levels.}

\item{y}{The label vector, positive(1) or negative(0)}

\item{n.comb}{The max number of features which are selected.}

\item{error.threshhold}{The threshhold of error.}

\item{fold}{The number of fold of cross validation used in the program.}

\item{device.id}{The GPU device ID.}

\item{cl}{The cluster of computers which you created. If it is NULL, the program will create clusters based on the number of devices automatically.}
}
\description{
Logistic Regression wiht CUDA
}

